refactor: extract ACP runtime handle ensure flow
